requirements.txt
# 核心深度学习框架
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# 科学计算基础
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0

# 机器学习和优化
scikit-learn>=1.3.0
optuna>=3.4.0

# 数据存储和处理
h5py>=3.11.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# 进度监控和日志
tqdm>=4.66.0
wandb>=0.16.0
tensorboard>=2.15.0

# 配置管理
PyYAML>=6.0
click>=8.1.0

# 测试和代码质量
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# 文档生成
Sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# 其他工具
joblib>=1.3.0
dill>=0.3.0